backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
praw>=7.7.0
requests>=2.31.0
transformers>=4.36.0
torch>=2.0.0
sentencepiece>=0.1.99
